Support for multiple python servers/virtual environments.
Individual transforms can either use a default python (i.e. in the PATH) or specify a path to a python executable. Additional path entries can be specified for the python environment to execute (e.g. under Windows Anaconda, .../Library/bin is an additional path requirement). An optional server ID can be also be supplied - this, in conjunction with the path to the python exe, is used to uniquely identify a python server instance. This allows a given server to be dedicated to one client, or shared among several.